+++ CMakeLists.txt
@@ -6,11 +6,11 @@ set(LIBPROC_SRC "/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/include
 set(CMAKE_OSX_DEPLOYMENT_TARGET "10.15" CACHE STRING "Minimum OS X deployment version")
 
 if (CMAKE_BUILD_TYPE MATCHES Debug)
-    message("Building debug")
-    add_compile_options(-Wall)
+    message("Building debug -o0")
+    add_compile_options(-Wall -o0)
 elseif(CMAKE_BUILD_TYPE MATCHES Release)
     message(STATUS "Building release")
-    add_compile_options(-Wall -gdwarf-2)
+    add_compile_options(-Wall -gdwarf-2 -o3)
 endif ()
 
 set(CMAKE_AUTOMOC ON)
