+++ src/calculate.h
@@ -8,7 +8,6 @@
 #include <string>
 #include "database.h"
 
-
 class Calculate {
 public:
     static double oz_alcohol(double abv, int amount);
@@ -19,8 +18,8 @@ public:
     static std::string favorite_producer(Storage storage);
     static std::string favorite_beer(Storage storage);
     static std::string favorite_type(Storage storage);
-    static double mean_abv(Storage storage);  // TODO write test
-    static double mean_ibu(Storage storage);  // TODO write test
+    static double mean_abv(Storage storage);
+    static double mean_ibu(Storage storage);
     static bool compare_date(const Drink &a, const Drink &b);
 };
