+++ test/test_database_functions.cpp
@@ -14,7 +14,7 @@ inline bool exists (const std::string& name) {
     return (stat (name.c_str(), &buffer) == 0);
 }
 
-TEST_CASE("Init DB", "[DB Creation]") {
+TEST_CASE("Init DB", "[DB Functions]") {
     std::string current_path = std::filesystem::current_path();
     const char *file_name = "testdb.db";
     std::string db_path = current_path + "/" + file_name;
@@ -29,7 +29,7 @@ TEST_CASE("Init DB", "[DB Creation]") {
     std::remove(file_name);
 }
 
-TEST_CASE("DB IO", "[DB IO]") {
+TEST_CASE("DB IO", "[DB Functions]") {
     std::string current_path = std::filesystem::current_path();
     const char *file_name = "testdb.db";
     std::string db_path = current_path + "/" + file_name;
@@ -65,7 +65,7 @@ TEST_CASE("DB IO", "[DB IO]") {
     REQUIRE(mosaic_read.type == "IPA");
 }
 
-TEST_CASE("Truncate DB", "[Truncate DB]") {
+TEST_CASE("Truncate DB", "[DB Functions]") {
     std::string current_path = std::filesystem::current_path();
     const char *file_name = "testdb.db";
     std::string db_path = current_path + "/" + file_name;
@@ -94,7 +94,7 @@ TEST_CASE("Truncate DB", "[Truncate DB]") {
     REQUIRE(storage_1.get_all<Drink>().empty() == true);
 }
 
-TEST_CASE("Delete Row", "[Delete Row]") {
+TEST_CASE("Delete Row", "[DB Functions]") {
     std::string current_path = std::filesystem::current_path();
     const char *file_name = "testdb.db";
     std::string db_path = current_path + "/" + file_name;
@@ -136,7 +136,7 @@ TEST_CASE("Delete Row", "[Delete Row]") {
     REQUIRE(max_id == min_id);
 }
 
-TEST_CASE("Read Row", "[Read Row]") {
+TEST_CASE("Read Row", "[DB Functions]") {
     std::string current_path = std::filesystem::current_path();
     const char *file_name = "testdb.db";
     std::string db_path = current_path + "/" + file_name;
@@ -168,7 +168,7 @@ TEST_CASE("Read Row", "[Read Row]") {
     REQUIRE(etrwo_read.notes == "Very good hazy IPA.");
 }
 
-TEST_CASE("Update Row", "[Update Row]") {
+TEST_CASE("Update Row", "[DB Functions]") {
     std::string current_path = std::filesystem::current_path();
     const char *file_name = "testdb.db";
     std::string db_path = current_path + "/" + file_name;
@@ -200,7 +200,7 @@ TEST_CASE("Update Row", "[Update Row]") {
     REQUIRE(etrwo_read2.notes == "Very good hazy IPA. Will buy again!");
 }
 
-TEST_CASE("Filter DB", "[Filter DB]") {
+TEST_CASE("Filter DB", "[DB Functions]") {
     std::string current_path = std::filesystem::current_path();
     const char *file_name = "testdb.db";
     std::string db_path = current_path + "/" + file_name;
@@ -233,7 +233,7 @@ TEST_CASE("Filter DB", "[Filter DB]") {
     REQUIRE(etrwo_read.drink_day == 8);
 }
 
-TEST_CASE("Get Drink By Name", "[DB Filter]") {
+TEST_CASE("Get Drink By Name", "[DB Functions]") {
     std::string current_path = std::filesystem::current_path();
     const char *file_name = "testdb.db";
     std::string db_path = current_path + "/" + file_name;
@@ -264,7 +264,7 @@ TEST_CASE("Get Drink By Name", "[DB Filter]") {
     REQUIRE(selected_beer.notes == "Very good hazy IPA.");
 }
 
-TEST_CASE("Get Beers By Type", "[DB Filter]") {
+TEST_CASE("Get Beers By Type", "[DB Functions]") {
     std::string current_path = std::filesystem::current_path();
     const char *file_name = "testdb.db";
     std::string db_path = current_path + "/" + file_name;
@@ -295,7 +295,7 @@ TEST_CASE("Get Beers By Type", "[DB Filter]") {
     REQUIRE(selected_beers.at(1).name == "Mosaic");
 }
 
-TEST_CASE("Get Beers By Brewery", "[DB Filter]") {
+TEST_CASE("Get Beers By Brewery", "[DB Functions]") {
     std::string current_path = std::filesystem::current_path();
     const char *file_name = "testdb.db";
     std::string db_path = current_path + "/" + file_name;
